Make private member names in keybind.cpp consistent
